requirements.txt
flask
pandas
scikit-learn
numpy
jinja2
matplotlib
seaborn
tensorflow
opencv-python
Pillow
